pyproject.toml
[project]
name = "insightforge"
version = "0.1.0"
description = "AI-powered market research and business analysis platform"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "anthropic>=0.45.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.18",
    "pydantic>=2.10.0",
    "python-dotenv>=1.0.0",
    "markdown>=3.7",
]
